http: Add two new exceptions, one constant
These will be useful in the future in case we don't enfore JSON encoding
anymore in the http.server module. The HTTP 1.1 RFC recommends error 415
(Unsupported Media Type) to be returned in case the client requests an
unsupported content-type. If the client doesn't send a “Content-Type” in
the request, a content-type of “application/octet-stream” is implied.

Signed-off-by: Michael Hanselmann <hansmi@google.com>
Reviewed-by: Iustin Pop <iustin@google.com>