install.rst: note about the default parameters
As we know, those are just defaults, mostly chosen by replicating the
only behavior we supported before allowing customization. They may need
changes, to work in specific environments (which is why we introduced
them in the first place).

This tries to address issue 83, which was caused by the default
parameters not being correct for the target cluster.

Signed-off-by: Guido Trotter <ultrotter@google.com>
Reviewed-by: Iustin Pop <iustin@google.com>